Feature/delete console

+++ client/src/pages/videopage2.js
@@ -2,23 +2,24 @@ import React, {useState, useEffect} from 'react';
 import { useHistory } from "react-router-dom";
 import axios from "axios"
 import './videopage2.css'
+import Comments from '../component/comments'
 
 axios.defaults.withCredentials = true;
 
-export default function VideoPage2({clickMyVideoData, userInfo, accessToken, viewStateFunc}){
-
+export default function VideoPage2({clickMyVideoData, userInfo, accessToken, viewStateFunc, isLogin, videoInfo}){
+  
     useEffect(() => {
         viewStateFunc(clickMyVideoData.id);
       }, []) 
+
+      
       
     const history = useHistory();
 
 
     // 업로드 영상 삭제하기
     const deleteMyVideoFunc = () => {
-        console.log(accessToken)
-        console.log(clickMyVideoData)
-        axios.delete(`https://localhost:80/myvideo?id=${clickMyVideoData.id}`,{
+        axios.delete(`${process.env.REACT_APP_SERVER}/myvideo?id=${clickMyVideoData.id}`,{
             headers: {
             authorization: `Bearer ${accessToken}`,
             "Content-Type" : "application/json"   
@@ -27,9 +28,8 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
           }
           )
           .then((res) => {
-              console.log(res)
               if(res.data.message === "deletion is complete") {
-                window.location.replace('/') 
+                window.location.replace('/main') 
               }
           })
           .catch((err) => {
@@ -39,7 +39,6 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
 
     const [likeVideo, setLikeVideo] = useState(false)
     const [likeVideoCount, setLikeVideoCount] = useState(clickMyVideoData.likes)
-
    // 좋아요 이모티콘  상태변경 함수
     const likeStateFunc = () => {
         setLikeVideo(!likeVideo);
@@ -55,8 +54,7 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
 
     // likeVideo가 false일 때 동영상 좋아요 post 요청
     const likeVideoFunc = () => {
-        console.log(likeVideo)
-        if(likeVideo === false) {
+        if(likeVideo === false && clickMyVideoData.mychoice === false) {
             axios.post("https://localhost:80/like/video",{video_id: clickMyVideoData.id},
             {
               headers: {
@@ -66,11 +64,10 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
               withCredentials: true
             })
             .then((res) => {
-                console.log(res)
                 likeStateFunc(); // 불이 켜지고
                 likeCountPlusFunc(); // 좋아요 카운트 +1
             })
-        } else if(likeVideo === true) {
+        } else if(likeVideo === false && clickMyVideoData.mychoice === true) {
             axios.delete(`https://localhost:80/like/video?video_id=${clickMyVideoData.id}`,
             {
               headers: {
@@ -80,9 +77,22 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
               withCredentials: true
             })
             .then((res) => {
-                console.log(res)
                 likeStateFunc(); // 불이 꺼지고
                 likeCountMinusFunc(); // 좋아요 카운트 -1
+            })
+        }
+        else if(likeVideo === true && clickMyVideoData.mychoice === true) {
+            axios.post("https://localhost:80/like/video",{video_id: clickMyVideoData.id},
+            {
+              headers: {
+              authorization: `Bearer ${accessToken}`,
+              "Content-Type" : "application/json"   
+              },
+              withCredentials: true
+            })
+            .then((res) => {
+                likeStateFunc(); // 불이 켜지고
+                likeCountPlusFunc(); // 좋아요 카운트 +1
             })
         }
     }
@@ -90,26 +100,37 @@ export default function VideoPage2({clickMyVideoData, userInfo, accessToken, vie
     
     return(
         <div className='videopage2-container'>
-            <video src={clickMyVideoData.video} width='80%' controls />
+            <div className="videopage2-videobox row-fluid">
+            <video className="videopage2-video col-md-9 col-11" src={clickMyVideoData.video} width='80%' controls />
+            </div>
             <div className='video2-info-container'>
                 <div className='video2-first-row'> 
+                  <div>
+                    <span className="video2-first-row-category">#{clickMyVideoData.category1}</span> 
+                    {clickMyVideoData.category2 ? <span className="video2-first-row-category">#{clickMyVideoData.category2}</span> : null}
+                    {clickMyVideoData.category3 ? <span className="video2-first-row-category">#{clickMyVideoData.category3}</span> : null}
+                  </div>
                     <div className='video2-first-row-title'>{clickMyVideoData.title}</div>
-                    <div className='video2-first-row-title'>{clickMyVideoData.createdAt}</div>
-                    <span>카테고리 {clickMyVideoData.category1}</span>
+                    <div className='video2-first-row-view-createdAt'>조회수 {clickMyVideoData.views} · {clickMyVideoData.createdAt.slice(0, 10)}</div>
                 </div> 
                 <div className='video2-second-row'>
-                    <div>작성자 {userInfo.name}</div>
-                    <div>view {clickMyVideoData.views}</div>
-                    <div onClick={likeVideoFunc}>찜
-                    {likeVideo === false ? 
-                    <img className="like-icon" src="https://i.ibb.co/C0ntKBK/2021-11-11-5-56-57-removebg-preview.png" alt="" /> :
-                    <img className="like-icon" src="https://i.ibb.co/Y0jmDXG/2021-11-11-5-58-41-removebg-preview.png" alt="" />
-                    } 
-                    {likeVideoCount}
+                    <div className='video2-second-row-creator'><img className="user-icon" src="https://i.ibb.co/ZV9MknX/profile-user.png"/> {userInfo.name}</div>
+                    <div className="video2-second-row-like-delete">
+                      <div className="like-myvideo btn2" onClick={likeVideoFunc}>
+                      {clickMyVideoData.mychoice === true && likeVideo === false ? 
+                      <img className="like-icon" src="https://i.ibb.co/Y0jmDXG/2021-11-11-5-58-41-removebg-preview.png" alt="" /> :
+                      <img className="like-icon" src="https://i.ibb.co/C0ntKBK/2021-11-11-5-56-57-removebg-preview.png" alt="" />
+                      }
+                      {likeVideoCount}
+                      </div>
+                      {isLogin === false ? null : 
+                      <div className="delete-myvideo btn2" onClick={deleteMyVideoFunc}>
+                      <img className="delete-icon" src="https://i.ibb.co/YkB7fMG/delete.png" /></div>
+                      }
                     </div>
-                    <div onClick={deleteMyVideoFunc}>삭제하기</div>
                 </div>
             </div>
+            <Comments accessToken={accessToken} videoInfo={videoInfo} userInfo={userInfo}/>
         </div>
     )
 }
